Add material ui theme
