+++ package.json
@@ -41,8 +41,9 @@
     "lint-staged": "^7.2.0",
     "prettier": "^1.14.0",
     "prop-types": "^15.6.2",
-    "react-styleguidist": "^7.1.0",
     "react-router-dom": "^4.3.1",
+    "react-styleguidist": "^7.1.0",
+    "storybook-addon-material-ui": "^0.9.0-alpha.18",
     "storybook-react-router": "^1.0.5"
   },
   "scripts": {
